Add classroom.rb  file
