+++ persons.rb
@@ -1,22 +1,37 @@
-class Person
+require_relative 'nameable'
+class Person < Nameable
+  attr_accessor :name, :age, :rentals
   attr_reader :id
 
-  attr_accessor :age, :name
-
-  def initialize(age, name = 'Unknown', parent_permission: true)
+  def initialize(age, name = 'unknown', parent_permission: true)
     @id = Random.rand(1..1000)
-    @age = age
     @name = name
+    @age = age.to_i
     @parent_permission = parent_permission
-  end
-
-  def can_use_services?
-    of_age? || @parent_permission
+    @rentals = []
+    super()
   end
 
   private
 
   def of_age?
-    @age >= 18
+    return true if @age >= 18
+
+    false
+  end
+
+  public
+
+  def can_use_services?
+    of_age? || @parent_permission
+  end
+
+  def correct_name
+    @name
+  end
+
+  def add_rental(rental)
+    @rentals.push(rental)
   end
 end
+
