+++ student.rb
@@ -1,13 +1,19 @@
 require_relative 'person'
-class Student < Person
-  attr_accessor :classroom
 
-  def initialize(id, age, classroom, name = 'unknown', parent_permission: true)
-    super(id, age, name, parent_permission)
-    @class_room = classroom
+class Student < Person
+  attr_reader :classroom
+
+  def initialize(age, classroom, name = 'unknown', parent_permission: true)
+    super(age, name, parent_permission: parent_permission)
+    @classroom = classroom
   end
 
   def play_hooky
     '¯(ツ)/¯'
+  end
+
+  def classroom=(classroom)
+    @classroom = classroom
+    classroom.students.push(self) unless classroom.students.include?(self)
   end
 end
