+++ teacher.rb
@@ -1,4 +1,4 @@
-require_relative 'persons'
+require_relative 'person'
 class Teacher < Person
   attr_accessor :specialization
